==> 00_libraries.py <==
#00_libraries.py
# This file lists the libraries required for the project.
#directories in all files must be changed! they are omitted for github version

"""
Project dependencies:

Core libraries:
    pip install --upgrade pip
    pip install duckdb polars pyarrow pandas pyreadstat 
    pip install joblib openai


Modeling (for later phases):
    pip install scikit-learn lightgbm xgboost 
    pip install shap matplotlib seaborn
"""
